[Refactor] OCR 라우터, 컨트롤러, 서비스 코드 분리
